\name{polarization}
\alias{polarization}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Polarization of allele frequencies
}
\description{
Determine, whether allele frequencies in a \code{sync} object are polarized for the minor, rising or reference allele.
}
\usage{
polarization(sync)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sync}{
a \code{sync} object.
}
}
\details{

}
\value{
A character indicating the polarization.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Thomas Taus
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{af}} and \code{\link{af.traj}}.
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
