\name{poolSeq-package}
\alias{poolSeq-package}
\alias{poolSeq}
\docType{package}
\title{
Analyze Pool-seq Data
}
\description{
The \code{poolSeq} package provides a variety of functions to simulate and analyze Pool-seq data, including estimation of the effective population size, selection coefficients and dominance parameters.
}
\details{

}
\author{
\packageAuthor{poolSeq}

Maintainer: \packageMaintainer{poolSeq}
}
\references{
Waples R. S.: A generalized approach for estimating effective population size from temporal changes in allele frequency, \emph{Genetics} \bold{1989}, 121, 379–391.

Agresti A.: Categorical data analysis (second edition). \emph{New York: Wiley} \bold{2002}

Jorde P. E. and Ryman N.: Unbiased estimator for genetic drift and effective population size, \emph{Genetics} \bold{2007}, 177 927–935.

Frick K., Munk, A. and Sieling, H.: Multiscale Change-Point Inference, \emph{Journal of the Royal Statistical Society: Series B} \bold{2014}, 76, 495-580.

Futschik A., Hotz T., Munk A. and Sieling H.: Multiresolution DNA partitioning: statistical evidence for segments, \emph{Bioinformatics} \bold{2014}, 30, 2255-2262.

Jónás A., Taus T., Kosiol C., Schlötterer C. & Futschik A.: Estimating effective population size from temporal allele frequency changes in experimental evolution, manuscript in preparation.
}
\keyword{ package }
\seealso{
\code{\link{wf.traj}}, \code{\link{sample.alleles}}, \code{\link{estimateNe}}, \code{\link{cmh.test}}, \code{\link{chi.sq.test}} and \code{\link{read.sync}}.
}
\examples{

}
